Extract workspace memory APIs
